requirements.txt
biopython==1.81  # Provides tools for biological computation